Factor out test helpers, add ava files config
